% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accident_2013.R
\docType{data}
\name{accident_2013}
\alias{accident_2013}
\title{Geographical Map depicting fatal injuries in mototr vehicle traffice crashes in United States}
\format{A data frame with 30202 rows and 50 variables:
\describe{
  \item{STATE}{State Number}
  \item{ST_CASE}{Consecutive Number}
  \item{VE_TOTAL}{Number of Vehicle Forms Submitted- ALL}
  \item{VE_FORMS}{Number of Motor Vehicles in Transport (MVIT) }
  \item{PVH_INVL}{Number of Parked/Working Vehicles}
  \item{PEDS}{Number of Forms Submitted for Persons Not in Motor Vehicles}
  \item{PERNOTMVIT}{Number of Persons Not in Motor Vehicles in Transport (MVIT)}
  \item{PERMVIT}{Number of Persons in Motor Vehicles in Transport (MVIT) }
  \item{PERSONS}{Number of Forms Submitted for Persons in Motor Vehicles}
  \item{COUNTY}{ County}
  \item{CITY}{ City}
  \item{DAY}{Day of crash}
  \item{MONTH}{month of crash}
  \item{YEAR}{year of Crash}
  \item{DAY_WEEK}{day of week}
  \item{HOUR}{hour of crash}
  \item{MINUTE}{minute of crash}
  \item{NHS}{National highway system}
  \item{ROAD_FNC}{functional road}
  \item{ROUTE}{Route signing}
  \item{TWAY_ID}{Traffic way identifier}
  \item{TWAY_ID2}{Traffic way identifier}
  \item{MILEPT}{mile point}
  \item{LATITUDE}{latitude}
  \item{LONGITUD}{longitude}
  \item{SP_JUR}{special jury}
  \item{HARM_EV}{first harmful event}
  \item{MAN_COLL}{manner of collision}
  \item{RELJCT2}{Relation to Junction- Within Interchange Area}
  \item{TYPE_INT}{type of intersection}
  \item{WRK_ZONE}{work zone}
  \item{REL_ROAD}{relation to trafficway}
  \item{LGT_COND}{light condition}
  \item{WEATHER1}{Atmospheric Conditions}
  \item{WEATHER2}{Atmospheric Conditions}
  \item{WEATHER}{Atmospheric Conditions}
  \item{SCH_BUS}{school bus}
  \item{RAIL}{Rail Grade Crossing Identifier}
  \item{NOT_HOUR}{hour of notificiation}
  \item{NOT_MIN}{minute of notification}
  \item{ARR_HOUR}{hour of arrival at scene}
  \item{ARR_MIN}{minute of arrival at scene}
  \item{HOSP_HR}{hour of ems arrival at hospital}
  \item{HOSP_MN}{minute of ems arrival at hospital}
  \item{CF1}{Related Factors- Crash Level }
  \item{CF2}{Related Factors- Crash Level }
  \item{CF3}{Related Factors- Crash Level }
  \item{FATALS}{fatality}
  \item{DRUNK_DR}{Number of Drinking Drivers}
}}
\source{
\url{https://crashstats.nhtsa.dot.gov/}
}
\usage{
accident_2013
}
\description{
Geographical Map depicting fatal injuries in mototr vehicle traffice crashes in United States
}
\keyword{datasets}
