% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Draws a geographical map for the accidental injuries for the given state and year.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state}{Number for each state. For more information , \href{Fatality Analysis Reporting System}{Fatality Analysis Reporting System}.}

\item{...}{Arguments passed on to \code{make_filename}
\describe{
  \item{year}{The year for which the information regarding fatal injuries is required. Valid value for <year>
will be those year for which user has "Accident_<year>.csv" file present in the local directory.}
}}
}
\value{
This function returns a geographical map for the given state(s) with each dot as an obsevation
of the accidental injuries.
}
\description{
This function reads the input file for the year provided by the user
and returns a geographical map depicting the fatal injuries suffered in motor vehicle
traffic crashes.
}
\details{
this function will throw an error if the year provided by user is not a valid year. For the definition on
valid year look at parameter 'year'.
}
\examples{
fars_map_state(1,2013)
fars_map_state(2,2015)
}
