% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Reading an input file.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{This can be either a variable holding the name of the file or the filename itself.}
}
\value{
This function returns a data frame for the input file.
}
\description{
This function checks whether a given file exists or not. If the file exists , it reads
the file into a data frame. Otherwise it throws an error with the message "File <filename>  does not exists".
}
\details{
Reads input file containing data regarding fatal injuries suffered in motor vehicle
traffic crashes.

If the system is not able to locate the given file in the current working
directory or the complete filename is not provided, this function will throw an error.
}
\examples{
fars_read("accident_2013.csv")
fars_read("accident_2015.csv")
data1<-"data/accident.csv"
fars_read(data1)
}
