% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Selecting and printing the months present in the input file for the given year.}
\usage{
fars_read_years(years)
}
\arguments{
\item{...}{Arguments passed on to \code{make_filename}
\describe{
  \item{year}{The year for which the information regarding fatal injuries is required. Valid value for <year>
will be those year for which user has "Accident_<year>.csv" file present in the local directory.}
}}
}
\value{
This function returns the year and the month present in the file for the selected year.
}
\description{
This function creates the filename by appending  user provided year to it.
It reads the file and selects the month present in the file for that year  as well as the year also.
}
\details{
This function will throw an error if the year provided by user is not a valid year.
}
\examples{
fars_read_years(2013)
fars_read_years(2015)
}
