% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Total number of accidental injuries based on month for the given year.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{...}{Arguments passed on to \code{make_filename}
\describe{
  \item{year}{The year for which the information regarding fatal injuries is required. Valid value for <year>
will be those year for which user has "Accident_<year>.csv" file present in the local directory.}
}}
}
\value{
This function returns the month and the total number of observation for each
 month present in the file for the given year.
}
\description{
This function reads the input file for the year provided by the user and
then returns the number of observation for each month of the given year.
}
\details{
this function will throw an error if the year provided by user is not a valid year.
}
\examples{
fars_summarize_years(2013)
fars_summarize_years(2015)
}
