% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Generating name of the input file based on year provided.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The year for which the information regarding fatal injuries is required. Valid value for <year>
will be those year for which user has "Accident_<year>.csv" file present in the local directory.}
}
\value{
This function returns the name of the file for the year provided.
}
\description{
This function returns the name of the file by formatting it with user supplied
year.
}
\examples{
make_filename(2013)
yr<-2013
make_filename(yr)
make_filename(2015)
}
